expose linkTools and default paperOptions for use outside of digitaljs
